+++ packages/react-resizable-panels-website/src/components/ResizeHandle.tsx
@@ -1,5 +1,6 @@
 import { PanelResizeHandle } from "react-resizable-panels";
 
+import Icon from "./Icon";
 import styles from "./ResizeHandle.module.css";
 
 export default function ResizeHandle({
@@ -14,7 +15,10 @@ export default function ResizeHandle({
       className={[styles.ResizeHandleOuter, className].join(" ")}
       id={id}
     >
-      <div className={styles.ResizeHandleInner} />
+      <div className={styles.ResizeHandleInner}>
+        <Icon className={styles.HorizontalIcon} type="resize-horizontal" />
+        <Icon className={styles.VerticalIcon} type="resize-vertical" />
+      </div>
     </PanelResizeHandle>
   );
 }