+++ packages/react-resizable-panels/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.19
+* Add optional `style` and `tagName` props to `Panel`, `PanelGroup`, and `PanelResizeHandle` to simplify custom styling.
+* Add `data-panel-group-direction` attribute to `PanelGroup` and `PanelResizeHandle` to simplify custom drag handle styling.
+
 ## 0.0.18
 * `Panel` and `PanelGroup` now use `overflow: hidden` style by default to avoid potential scrollbar flickers while resizing.